Update swiper language to represent carousels/sliders
