+++ resources/assets/scripts/lib/swipers.js
@@ -1,0 +1,22 @@
+import SwiperCore, { Navigation, Pagination } from 'swiper/core';
+SwiperCore.use([Navigation, Pagination]);
+
+export const makeSwiper = (el, options = {}) => {
+  const defaults = {
+    pagination: {
+      el: '.slab-swiper__pagination',
+    },
+    navigation: {
+      prevEl: '.slab-swiper__button-prev',
+      nextEl: '.slab-swiper__button-next',
+    },
+  };
+
+  return new SwiperCore(el, Object.assign({}, defaults, options));
+};
+
+export const initSliders = () => {
+  const swiperEls = document.querySelectorAll('.slab-swiper--slider');
+
+  Array.prototype.forEach.call(swiperEls, makeSwiper);
+};
