+++ resources/views/components/slide.blade.php
@@ -1,8 +1,8 @@
 <div class="swiper-slide slab-slide">
   <div class="slab-slide__bg">
-    <img data-src="{{ wp_get_attachment_image_src($background_image, 'large')[0] }}" class="swiper-lazy" />
-    <div class="swiper-lazy-preloader swiper-lazy-preloader-white"></div>
+    {!! wp_get_attachment_image($background_image, 'large') !!}
   </div>
+  <div class="slab-slide__overlay"></div>
   <div class="slab-slide__content">
     <h3>{{ $title }}</h3>
     <p>{{ $description }}</p>
